adds ability to rename scenario
