+++ app/layout/scenarios/sidebar/component.tsx
@@ -3,16 +3,13 @@ import React, { PropsWithChildren } from 'react';
 import { AnimatePresence } from 'framer-motion';
 
 import Breadcrumbs from 'layout/project/navigation/breadcrumbs';
-import { cn } from 'utils/cn';
+import ScenarioHeader from 'layout/project/sidebar/scenario/header';
 
 export const ScenariosEditSidebar = ({ children }: PropsWithChildren): JSX.Element => {
   return (
-    <div
-      className={cn({
-        'flex w-full flex-grow flex-col overflow-hidden': true,
-      })}
-    >
+    <div className="flex w-full flex-grow flex-col overflow-hidden">
       <Breadcrumbs />
+      <ScenarioHeader />
       <div className="mt-2.5 flex flex-grow flex-col overflow-hidden">
         <AnimatePresence>{children}</AnimatePresence>
       </div>
